Add code for the color picker feature and fix bugs
just added code for controlling the new feature which is simply a color picker. I have also fixed some bugs which i did not notice before, and these bugs were concerned with the fact that when the grid size was increased all the number of boxes to draw on increased, the newly added boxes were not being responsive.