Edit style.css
Just Edited the CSS file so as to add styling code for the color picker. And i have also added  a class selector called "#enabled-btn-effect". This class selector is supposed to help me with refactoring my JavaScript code by just adding and removing this class from elements in order to show that they are either active or not